Fix deadlock in CPU thread
Finally found the culprit causing that "emulation thread failed to
terminate plugins" error. It had nothing to do with plugins! The
deadlock is caused by the CPU thread sending messages to the GUI
thread, ALL INTERNAL to PJ64.

When this happens, the two threads wait on each other and no progress
can be made. Deadlock!

The proper fix is to make the GUI thread continue processing messages
while it waits for the CPU thread to stop.

This does keep the disaster recovery code, even though we do not want it
to be called, ever. It is probably best to just remove it. The new upper
bound deadline for disaster recovery is 2 seconds. Which should be
plenty of time for the threads to cooperate, even on heavily loaded (or
very slow) systems.