Fix build problems (#51)
I was trying to find a way to fix the "file already in use" error when building for release. This
happens when the manifest is embedded into the exe. StackOverflow said AV software was responsible.
I tried some other workarounds but they all failed. In the end I had to add a local exception for my
build directory to Windows Defender! Keep up the good work, MS.

On the other hand, I did fix that weird warning about the trailing path separator, and made some
minor cleanups to the project file. So, we got that out of it.