Fix x86 register cache in DSUBU and others
- The register protections mark native registers so that they are not
  reused for allocation when there are no free registers. This is needed
  for any "source" MIPS registers to avoid clobbering when mapping MIPS
  registers to x86.

- There are a very limited number of x86 registers (just 10!) but only
  up to 4 need to be protected at any given time; 2 64-bit MIPS source
  registers = 4 x86 registers. (I do not believe there are any MIPS
  instructions that require sourcing from more than 2 registers.)

- All protections are removed after every MIPS instruction is compiled.
  There is usually no need to unprotect the register before returning.

The register protection was completely missing in DSUBU (the wrong
register was protected later than it should have been, anyway). It needs
to protect rd before mapping a temporary to rt. And rt needs to be
protected before rd is mapped.

I have attempted to fix all of the missing protections that I could
find. There might be others... DSUBU was enough to fix part of Stunt
Racer 64.