Memory-mapped ROM file
This saves a lot of heap space by memory mapping the ROM (the memory is
file-backed) instead of retaining the entire file contents on the heap.

It has to create a temporary file because it can't map compressed data
(from a zip file) and the entire emulator and plugin ecosystem expects
the ROM to be stored in little endian. There's also a little hidden
hack for Aleck64 games that patches the ROM contents...

As a bonus, the memory mapped file is always read-only. See
`FILE_MAP_READ` documentation in
https://learn.microsoft.com/en-us/windows/win32/api/memoryapi/nf-memoryapi-mapviewoffile
This allows removing several VirtualProtect() calls.