Refactor watchpoints implementation for maintainability
This replaces the random-looking bit twiddling with comments and
variable names that better illustrate the purpose of the twiddling.

I came back to this code a year-ish after writing it and it was
mystifying! I don't use C++ enough to remember the details of the
unordered_map API, or why it indexes the search result twice with a
bunch of bit masking on both index operations. Hopefully this helps
future-me understand what the heck past-me was thinking.

I also note here in a TODO comment that the public HasWatchPoint API is
used internally by ToggleWatchPoint in a dubious way. This only works as
long as the watchpoint slot doesn't need to contain any more information
than what the WATCH_TYPE enum carries. (And in at least one intermediate
phase, that certainly wasn't the case! See
https://github.com/pj64team/Project64-Legacy/pull/54 force-commit
history where I used to store the upper-bits of the 32-bit address in
the slot. Whoops!)